feat: create separate component Search
