+++ src/pages/home/search/search.scss
@@ -1,0 +1,81 @@
+@import "../../../styles/constants";
+@import "../../../styles/mixins";
+
+.search {
+  width: 500px;
+  position: relative;
+
+  @include media-md {
+    width: 350px;
+  }
+
+  @include media-sm {
+    width: 300px;
+  }
+}
+
+.search__input {
+  padding: 10px 40px;
+  width: 100%;
+  font-size: map-get($font-size, l);
+  font-weight: 500;
+  color: $color-font-dark-regular;
+  border: 2px solid $color-font-dark-regular;
+  border-radius: $border-radius-md;
+  box-sizing: border-box;
+  transition: $transition;
+  outline: none;
+
+  &:active,
+  &:focus,
+  &:hover {
+    border-color: $color-accent;
+  }
+
+  @include media-sm {
+    font-size: map-get($font-size, m);
+  }
+}
+
+.search__icon {
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+
+  &:hover {
+    .search__input {
+      border-color: $color-accent;
+    }
+  }
+
+  &--search {
+    @include icon-background(24px, 24px, "../../../assets/images/search.svg");
+    left: 10px;
+  }
+}
+
+.search__icon--arrow {
+  @include icon-background(24px, 24px, "../../../assets/images/arrow.svg");
+}
+
+.search__button {
+  width: 24px;
+  height: 24px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: absolute;
+  top: 50%;
+  right: 10px;
+  transform: translateY(-50%);
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+  outline: none;
+
+  &:hover {
+    .search__input {
+      border-color: $color-accent;
+    }
+  }
+}
